--- a/central/processbaseline/evaluator/evaluator_test.go
+++ b/central/processbaseline/evaluator/evaluator_test.go
@@ -1,6 +1,7 @@
 package evaluator
 
 import (
+	"context"
 	"strings"
 	"testing"
 	"time"
@@ -9,6 +10,7 @@ import (
 	processBaselineResultMocks "github.com/stackrox/rox/central/processbaselineresults/datastore/mocks"
 	processIndicatorMocks "github.com/stackrox/rox/central/processindicator/datastore/mocks"
 	"github.com/stackrox/rox/central/processindicator/views"
+	v1 "github.com/stackrox/rox/generated/api/v1"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/fixtures"
 	"github.com/stackrox/rox/pkg/protoconv"
@@ -243,7 +245,13 @@ func TestProcessBaselineEvaluator(t *testing.T) {
 
 			mockBaselines.EXPECT().GetProcessBaseline(gomock.Any(), gomock.Any()).MaxTimes(len(deployment.GetContainers())).Return(c.baseline, c.baseline != nil, c.baselineErr)
 			if c.indicators != nil {
-				mockIndicators.EXPECT().GetProcessIndicatorsRiskView(gomock.Any(), gomock.Any()).Return(c.indicators, c.indicatorErr)
+				mockIndicators.EXPECT().IterateOverProcessIndicatorsRiskView(gomock.Any(), gomock.Any(), gomock.Any()).Do(
+					func(_ context.Context, _ *v1.Query, fn func(indicator *views.ProcessIndicatorRiskView) error) {
+						for _, i := range c.indicators {
+							err := fn(i)
+							require.NoError(t, err)
+						}
+					}).Return(c.indicatorErr)
 			}
 
 			expectedBaselineResult := &storage.ProcessBaselineResults{