The current `GetProcessIndicatorsRiskView` calls are inefficient for our use case, since we need to loop through the returned process indicators. Switching to a callback-based approach allows us to process items one by one, avoiding the need to load all objects into a slice. This reduces memory usage and lowers GC pressure.


```bash
go test -tags sql_integration -run=NO -bench=BenchmarkEvaluateBaselinesAndPersistResult  ./central/processbaseline/evaluator/. -count 6 -benchmem -benchtime 3s 
```

```
                                                                              │ master-2.txt │               new.txt               │
                                                                              │    sec/op    │    sec/op     vs base               │
EvaluateBaselinesAndPersistResult/100_processes_2_containers-8                  1.781m ± 32%   1.720m ± 55%        ~ (p=0.240 n=6)
EvaluateBaselinesAndPersistResult/500_processes_3_containers-8                  4.968m ± 28%   4.865m ± 43%        ~ (p=0.485 n=6)
EvaluateBaselinesAndPersistResult/1000_processes_5_containers-8                 9.076m ± 46%   8.471m ± 32%        ~ (p=0.180 n=6)
EvaluateBaselinesAndPersistResult/1000_processes_5_containers_large_args-8      11.86m ± 12%   11.64m ±  9%        ~ (p=0.485 n=6)
EvaluateBaselinesAndPersistResult/2000_processes_10_containers-8                15.86m ± 29%   15.19m ± 29%        ~ (p=0.240 n=6)
EvaluateBaselinesAndPersistResult/10000_processes_20_containers-8               81.46m ± 44%   74.28m ± 17%        ~ (p=0.132 n=6)
EvaluateBaselinesAndPersistResult/25000_processes_50_containers-8               228.4m ± 14%   182.5m ± 12%  -20.12% (p=0.026 n=6)
EvaluateBaselinesAndPersistResult/50000_processes_100_containers-8              483.2m ± 17%   346.9m ± 33%  -28.21% (p=0.009 n=6)
EvaluateBaselinesAndPersistResult/50000_processes_100_containers_large_args-8   633.3m ± 24%   542.8m ± 13%  -14.30% (p=0.004 n=6)
geomean                                                                         35.34m         31.66m        -10.41%

                                                                              │  master-2.txt  │                new.txt                │
                                                                              │ max_heap_bytes │ max_heap_bytes  vs base               │
EvaluateBaselinesAndPersistResult/100_processes_2_containers-8                    11.23M ±  9%     11.23M ± 17%        ~ (p=0.818 n=6)
EvaluateBaselinesAndPersistResult/500_processes_3_containers-8                    13.56M ± 18%     12.79M ± 15%        ~ (p=0.132 n=6)
EvaluateBaselinesAndPersistResult/1000_processes_5_containers-8                   15.10M ± 13%     13.50M ± 28%        ~ (p=0.240 n=6)
EvaluateBaselinesAndPersistResult/1000_processes_5_containers_large_args-8        20.17M ± 35%     18.61M ± 33%   -7.73% (p=0.041 n=6)
EvaluateBaselinesAndPersistResult/2000_processes_10_containers-8                  16.49M ±  7%     15.79M ±  2%   -4.28% (p=0.004 n=6)
EvaluateBaselinesAndPersistResult/10000_processes_20_containers-8                 38.74M ±  9%     32.44M ±  8%  -16.27% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/25000_processes_50_containers-8                 76.33M ±  5%     39.86M ± 18%  -47.78% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/50000_processes_100_containers-8               137.26M ±  3%     41.16M ± 19%  -70.02% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/50000_processes_100_containers_large_args-8    362.78M ±  6%     86.84M ± 35%  -76.06% (p=0.002 n=6)
geomean                                                                           36.42M           23.99M        -34.12%

                                                                              │   master-2.txt   │                 new.txt                 │
                                                                              │ max_heap_objects │ max_heap_objects  vs base               │
EvaluateBaselinesAndPersistResult/100_processes_2_containers-8                      11.23M ±  9%       11.23M ± 17%        ~ (p=0.818 n=6)
EvaluateBaselinesAndPersistResult/500_processes_3_containers-8                      13.56M ± 18%       12.79M ± 15%        ~ (p=0.132 n=6)
EvaluateBaselinesAndPersistResult/1000_processes_5_containers-8                     15.10M ± 13%       13.50M ± 28%        ~ (p=0.240 n=6)
EvaluateBaselinesAndPersistResult/1000_processes_5_containers_large_args-8          20.17M ± 35%       18.61M ± 33%   -7.73% (p=0.041 n=6)
EvaluateBaselinesAndPersistResult/2000_processes_10_containers-8                    16.49M ±  7%       15.79M ±  2%   -4.28% (p=0.004 n=6)
EvaluateBaselinesAndPersistResult/10000_processes_20_containers-8                   38.74M ±  9%       32.44M ±  8%  -16.27% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/25000_processes_50_containers-8                   76.33M ±  5%       39.86M ± 18%  -47.78% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/50000_processes_100_containers-8                 137.26M ±  3%       41.16M ± 19%  -70.02% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/50000_processes_100_containers_large_args-8      362.78M ±  6%       86.84M ± 35%  -76.06% (p=0.002 n=6)
geomean                                                                             36.42M             23.99M        -34.12%

                                                                              │ master-2.txt │               new.txt               │
                                                                              │     B/op     │     B/op      vs base               │
EvaluateBaselinesAndPersistResult/100_processes_2_containers-8                  678.9Ki ± 0%   660.3Ki ± 0%   -2.74% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/500_processes_3_containers-8                  2.116Mi ± 0%   2.029Mi ± 0%   -4.11% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/1000_processes_5_containers-8                 3.774Mi ± 0%   3.609Mi ± 0%   -4.39% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/1000_processes_5_containers_large_args-8      8.955Mi ± 0%   8.790Mi ± 0%   -1.85% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/2000_processes_10_containers-8                7.132Mi ± 0%   6.673Mi ± 0%   -6.43% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/10000_processes_20_containers-8               34.66Mi ± 0%   31.69Mi ± 0%   -8.55% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/25000_processes_50_containers-8               87.49Mi ± 0%   77.93Mi ± 0%  -10.93% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/50000_processes_100_containers-8              175.7Mi ± 1%   155.0Mi ± 0%  -11.80% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/50000_processes_100_containers_large_args-8   463.0Mi ± 0%   442.3Mi ± 0%   -4.47% (p=0.002 n=6)
geomean                                                                         16.35Mi        15.33Mi        -6.20%

                                                                              │ master-2.txt │              new.txt              │
                                                                              │  allocs/op   │  allocs/op   vs base              │
EvaluateBaselinesAndPersistResult/100_processes_2_containers-8                   15.59k ± 0%   15.45k ± 0%  -0.90% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/500_processes_3_containers-8                   57.99k ± 0%   57.81k ± 0%  -0.31% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/1000_processes_5_containers-8                  108.7k ± 0%   108.5k ± 0%  -0.18% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/1000_processes_5_containers_large_args-8       108.7k ± 0%   108.5k ± 0%  -0.18% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/2000_processes_10_containers-8                 208.7k ± 0%   208.4k ± 0%  -0.11% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/10000_processes_20_containers-8                1.015M ± 0%   1.015M ± 0%  -0.03% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/25000_processes_50_containers-8                2.515M ± 0%   2.515M ± 0%  -0.02% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/50000_processes_100_containers-8               5.015M ± 0%   5.015M ± 0%  -0.01% (p=0.002 n=6)
EvaluateBaselinesAndPersistResult/50000_processes_100_containers_large_args-8    5.015M ± 0%   5.015M ± 0%  -0.01% (p=0.002 n=6)
geomean                                                                          374.0k        373.2k       -0.20%
```

refs:
- https://github.com/stackrox/stackrox/pull/15085
- https://github.com/stackrox/stackrox/pull/15010
- #17078 
- https://github.com/stackrox/stackrox/pull/16965
- https://github.com/stackrox/stackrox/pull/16987